flight_pattern_of_life-feature-data-extractor/requirements.txt
tensorboard
lightning
torch 
torchvision 
torchaudio
folium
selenium
hydra-core 
pytorch-lightning 
omegaconf 
scipy 
matplotlib 
numpy 
pandas 
mlflow
kfp==2.9.0
boto3